+++ libs/maps/include/mrpt/maps/CVoxelMapOccupancyBase.h
@@ -35,6 +35,10 @@ struct TVoxelMap_InsertionOptions : public mrpt::config::CLoadableOptions
 	bool ray_trace_free_space = true;
 	uint32_t decimation = 1;
 
+	/** If !=0, remove the voxels farther (L1 distance) than this
+	 * distance, in meters. */
+	double remove_voxels_farther_than = .0;
+
 	// See base docs
 	void loadFromConfigFile(
 		const mrpt::config::CConfigFileBase& source,
@@ -409,9 +413,10 @@ void CVoxelMapOccupancyBase<voxel_node_t, occupancy_t>::getAsOctoMapVoxels(
 	const double bbox_span_z_inv = 1.0 / bbox_span_z;
 
 	// Go thru all voxels:
-	auto lmbdPerVoxel = [this, &grid, &gl_obj, general_color_u, general_color,
-						 bbox, bbox_span_z_inv](
-							voxel_node_t& data, const Bonxai::CoordT& coord) {
+	auto lmbdPerVoxel =
+		[this, &grid, &gl_obj, general_color_u, general_color, bbox,
+		 bbox_span_z_inv](voxel_node_t& data, const Bonxai::CoordT& coord)
+	{
 		using mrpt::img::TColor;
 
 		// log-odds to probability:
@@ -711,7 +716,8 @@ void CVoxelMapOccupancyBase<voxel_node_t, occupancy_t>::updateCachedProperties()
 
 	// Go thru all voxels:
 	auto lmbdPerVoxel = [this, freenessThreshold, &grid](
-							voxel_node_t& data, const Bonxai::CoordT& coord) {
+							voxel_node_t& data, const Bonxai::CoordT& coord)
+	{
 		using mrpt::img::TColor;
 
 		// log-odds to probability: