+++ libs/maps/src/maps/CVoxelMapOccupancyBase.cpp
@@ -23,6 +23,7 @@ void TVoxelMap_InsertionOptions::loadFromConfigFile(
 	MRPT_LOAD_CONFIG_VAR(clamp_max, double, c, s);
 	MRPT_LOAD_CONFIG_VAR(ray_trace_free_space, bool, c, s);
 	MRPT_LOAD_CONFIG_VAR(decimation, uint64_t, c, s);
+	MRPT_LOAD_CONFIG_VAR(remove_voxels_farther_than, double, c, s);
 }
 void TVoxelMap_InsertionOptions::saveToConfigFile(
 	mrpt::config::CConfigFileBase& c, const std::string& s) const
@@ -34,16 +35,18 @@ void TVoxelMap_InsertionOptions::saveToConfigFile(
 	MRPT_SAVE_CONFIG_VAR(clamp_max, c, s);
 	MRPT_SAVE_CONFIG_VAR(ray_trace_free_space, c, s);
 	MRPT_SAVE_CONFIG_VAR(decimation, c, s);
+	MRPT_SAVE_CONFIG_VAR(remove_voxels_farther_than, c, s);
 }
 
 void TVoxelMap_InsertionOptions::writeToStream(
 	mrpt::serialization::CArchive& out) const
 {
-	const uint8_t version = 0;
+	const uint8_t version = 1;
 	out << version;
 
 	out << max_range << prob_miss << prob_hit << clamp_min << clamp_max;
 	out << ray_trace_free_space << decimation;
+	out << remove_voxels_farther_than;	// v1
 }
 
 void TVoxelMap_InsertionOptions::readFromStream(
@@ -53,8 +56,10 @@ void TVoxelMap_InsertionOptions::readFromStream(
 	switch (version)
 	{
 		case 0:
+		case 1:
 			in >> max_range >> prob_miss >> prob_hit >> clamp_min >> clamp_max;
 			in >> ray_trace_free_space >> decimation;
+			if (version >= 1) { in >> remove_voxels_farther_than; }
 			break;
 		default: MRPT_THROW_UNKNOWN_SERIALIZATION_VERSION(version);
 	}